♻️ refactor : 빌드오류해결 suspense 처리
